+++ packages/framework/tests/Unit/IncludesFacadeUnitTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit;
 
+use Illuminate\Contracts\View\Factory;
 use Mockery;
 use Closure;
 use Hyde\Hyde;
@@ -150,6 +151,8 @@ class IncludesFacadeUnitTest extends UnitTestCase
             $filesystem->shouldReceive('get')->with($this->includesPath($filename))->andReturn($content);
         });
 
+        $this->mockViewFactory($expected);
+
         $this->assertHtmlStringIsSame($expected, Includes::markdown($filename));
     }
 
@@ -162,6 +165,8 @@ class IncludesFacadeUnitTest extends UnitTestCase
         $this->mockFilesystemFromClosure(function ($filesystem) use ($filename) {
             $filesystem->shouldReceive('exists')->with($this->includesPath($filename))->andReturn(false);
         });
+
+        $this->mockViewFactory($expected);
 
         $this->assertNull(Includes::markdown($filename));
         $this->assertHtmlStringIsSame($expected, Includes::markdown($filename, $default));
@@ -178,6 +183,8 @@ class IncludesFacadeUnitTest extends UnitTestCase
             $filesystem->shouldReceive('exists')->with($this->includesPath($filename))->andReturn(true);
             $filesystem->shouldReceive('get')->with($this->includesPath($filename))->andReturn($content);
         });
+
+        $this->mockViewFactory($expected);
 
         $this->assertHtmlStringIsSame($expected, Includes::markdown('foo.md'));
         $this->assertHtmlStringIsSame(Includes::markdown('foo.md'), Includes::markdown('foo'));
@@ -240,6 +247,18 @@ class IncludesFacadeUnitTest extends UnitTestCase
         $config($filesystem);
 
         app()->instance(Filesystem::class, $filesystem);
+    }
+
+    protected function mockViewFactory(string $result): void
+    {
+        $viewFactory = Mockery::mock(Factory::class);
+        $viewFactory->shouldReceive('make')
+            ->andReturn(Mockery::mock([
+                'render' => $result,
+            ]))->byDefault();
+
+        app()->instance('view', $viewFactory);
+        app()->instance(Factory::class, $viewFactory);
     }
 
     protected function includesPath(string $filename): string